refactor(api): split main.py (1264 LOC) into resource routers (#71)
Closes Phase 3's first issue. main.py: 1264 → 181 LOC.

Pure relocation refactor. No route URL changes; no behavior changes;
no auth/permission changes. The post-#69/#70 dataclass and auth
Depends are preserved verbatim.

NEW files (8):
- core/api/search.py (181 LOC) — 3 search endpoints
  (fulltext/semantic/hybrid)
- core/api/pruning.py (128 LOC) — POST /memory/prune/{suggest,confirm}
- core/api/compression.py (126 LOC) — POST /memory-blocks/{id}/compress
  and /compress/apply
- core/api/user_info.py (165 LOC) — GET /user-info (kept its own router
  since it has no /users prefix)
- core/api/memory_blocks_bulk.py (376 LOC) — POST /memory-blocks/bulk-*
  (3 routes; standalone router with full paths to avoid wildcard
  ordering conflicts with /memory-blocks/{id})
- core/api/middleware/__init__.py + scope.py (38 LOC) —
  enforce_write_scope_metadata middleware
- core/services/keyword_extraction_service.py (75 LOC) —
  extract_keywords_enhanced helper that bulk-keywords endpoints use

MOVED to existing files:
- POST /memory-blocks/{id}/change-scope → memory_blocks.py (matches
  existing /memory-blocks prefix)

KEPT in main.py:
- GET /health (app-level)
- GET /conversations/count — moved back to main.py with @app.get
  after initial agent attempt put it in memory_blocks.py where it
  would have resolved at /memory-blocks/conversations/count
  (URL-breaking)

main.py contents (181 LOC):
- App factory + CORS + identity-mismatch exception handler
- enforce_readonly_for_guests middleware (kept inline; small)
- enforce_write_scope_metadata middleware (registered, defined in
  middleware/scope.py)
- Router registration (16 routers)
- /health and /conversations/count

Test fix bundled:
- tests/integration/memory_blocks/test_memory_optimization.py: patch
  target updated from core.api.main.extract_keywords_enhanced to
  core.services.keyword_extraction_service.extract_keywords_enhanced

Net diff: 11 files changed, +1269/-1136 (net +133 LOC; the increase is
the per-file router boilerplate — necessary cost of the split).

Verification (run locally; integration branch is CI-silent until
staging PR per .github/workflows/deploy.yml):
- uv run pytest (full suite): 859 passed, 8 skipped, 80.31% coverage
- bun run build (Phase 3 cross-stack sanity): clean, 2.56s
- main.py route count: 2 (/health, /conversations/count) — down from 14
- All moved route URLs verified unchanged via the test suite (no test
  needed URL updates beyond the one mock-target fix)

Phase 3 starts: #71 done, #72-#75 pending (crud → repositories,
notification_service split, orgs split, search_service Strategy).

Refs #71, umbrella #85.
